chore: refactor into payload.rs and engine.rs
