+++ src/server.rs
@@ -2,6 +2,11 @@ use crate::{
     HealthHandle,
     client::rpc::RpcClient,
     debug_api::DebugServer,
+    engine::EngineApiServer,
+    payload::{
+        NewPayload, NewPayloadV3, NewPayloadV4, OpExecutionPayloadEnvelope, PayloadSource,
+        PayloadTraceContext, PayloadVersion,
+    },
     probe::{Health, Probes},
 };
 use alloy_primitives::{B256, Bytes};
@@ -29,97 +34,6 @@ use op_alloy_rpc_types_engine::{
 };
 use tokio::task::JoinHandle;
 use tracing::{info, instrument};
-
-const CACHE_SIZE: u64 = 100;
-
-#[derive(Debug, Clone)]
-pub struct PayloadTrace {
-    pub builder_has_payload: bool,
-    pub trace_id: Option<tracing::Id>,
-}
-
-pub struct PayloadTraceContext {
-    block_hash_to_payload_ids: Cache<B256, Vec<PayloadId>>,
-    payload_id: Cache<PayloadId, PayloadTrace>,
-}
-
-impl PayloadTraceContext {
-    fn new() -> Self {
-        PayloadTraceContext {
-            block_hash_to_payload_ids: Cache::new(CACHE_SIZE),
-            payload_id: Cache::new(CACHE_SIZE),
-        }
-    }
-
-    async fn store(
-        &self,
-        payload_id: PayloadId,
-        parent_hash: B256,
-        builder_has_payload: bool,
-        trace_id: Option<tracing::Id>,
-    ) {
-        self.payload_id
-            .insert(
-                payload_id,
-                PayloadTrace {
-                    builder_has_payload,
-                    trace_id,
-                },
-            )
-            .await;
-        self.block_hash_to_payload_ids
-            .entry(parent_hash)
-            .and_upsert_with(|o| match o {
-                Some(e) => {
-                    let mut payloads = e.into_value();
-                    payloads.push(payload_id);
-                    std::future::ready(payloads)
-                }
-                None => std::future::ready(vec![payload_id]),
-            })
-            .await;
-    }
-
-    async fn trace_ids_from_parent_hash(&self, parent_hash: &B256) -> Option<Vec<tracing::Id>> {
-        match self.block_hash_to_payload_ids.get(parent_hash).await {
-            Some(payload_ids) => Some(
-                stream::iter(payload_ids.iter())
-                    .filter_map(|payload_id| async {
-                        self.payload_id
-                            .get(payload_id)
-                            .await
-                            .and_then(|x| x.trace_id)
-                    })
-                    .collect()
-                    .await,
-            ),
-            None => None,
-        }
-    }
-
-    async fn trace_id(&self, payload_id: &PayloadId) -> Option<tracing::Id> {
-        self.payload_id
-            .get(payload_id)
-            .await
-            .and_then(|x| x.trace_id)
-    }
-
-    async fn has_builder_payload(&self, payload_id: &PayloadId) -> bool {
-        self.payload_id
-            .get(payload_id)
-            .await
-            .map(|x| x.builder_has_payload)
-            .unwrap_or_default()
-    }
-
-    async fn remove_by_parent_hash(&self, block_hash: &B256) {
-        if let Some(payload_ids) = self.block_hash_to_payload_ids.remove(block_hash).await {
-            for payload_id in payload_ids.iter() {
-                self.payload_id.remove(payload_id).await;
-            }
-        }
-    }
-}
 
 pub struct RollupBoostServer {
     pub l2_client: Arc<RpcClient>,
@@ -185,74 +99,6 @@ impl TryInto<RpcModule<()>> for RollupBoostServer {
 
         Ok(module)
     }
-}
-
-#[derive(Debug, Clone)]
-pub enum PayloadSource {
-    L2,
-    Builder,
-}
-
-impl std::fmt::Display for PayloadSource {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            PayloadSource::L2 => write!(f, "l2"),
-            PayloadSource::Builder => write!(f, "builder"),
-        }
-    }
-}
-
-#[allow(dead_code)]
-impl PayloadSource {
-    pub fn is_builder(&self) -> bool {
-        matches!(self, PayloadSource::Builder)
-    }
-
-    pub fn is_l2(&self) -> bool {
-        matches!(self, PayloadSource::L2)
-    }
-}
-
-#[rpc(server, client)]
-pub trait EngineApi {
-    #[method(name = "engine_forkchoiceUpdatedV3")]
-    async fn fork_choice_updated_v3(
-        &self,
-        fork_choice_state: ForkchoiceState,
-        payload_attributes: Option<OpPayloadAttributes>,
-    ) -> RpcResult<ForkchoiceUpdated>;
-
-    #[method(name = "engine_getPayloadV3")]
-    async fn get_payload_v3(
-        &self,
-        payload_id: PayloadId,
-    ) -> RpcResult<OpExecutionPayloadEnvelopeV3>;
-
-    #[method(name = "engine_newPayloadV3")]
-    async fn new_payload_v3(
-        &self,
-        payload: ExecutionPayloadV3,
-        versioned_hashes: Vec<B256>,
-        parent_beacon_block_root: B256,
-    ) -> RpcResult<PayloadStatus>;
-
-    #[method(name = "engine_getPayloadV4")]
-    async fn get_payload_v4(
-        &self,
-        payload_id: PayloadId,
-    ) -> RpcResult<OpExecutionPayloadEnvelopeV4>;
-
-    #[method(name = "engine_newPayloadV4")]
-    async fn new_payload_v4(
-        &self,
-        payload: OpExecutionPayloadV4,
-        versioned_hashes: Vec<B256>,
-        parent_beacon_block_root: B256,
-        execution_requests: Vec<Bytes>,
-    ) -> RpcResult<PayloadStatus>;
-
-    #[method(name = "eth_getBlockByNumber")]
-    async fn get_block_by_number(&self, number: BlockNumberOrTag, full: bool) -> RpcResult<Block>;
 }
 
 #[async_trait]
@@ -369,7 +215,7 @@ impl EngineApiServer for RollupBoostServer {
     ) -> RpcResult<OpExecutionPayloadEnvelopeV3> {
         info!("received get_payload_v3");
 
-        match self.get_payload(payload_id, Version::V3).await? {
+        match self.get_payload(payload_id, PayloadVersion::V3).await? {
             OpExecutionPayloadEnvelope::V3(v3) => Ok(v3),
             OpExecutionPayloadEnvelope::V4(_) => Err(ErrorObject::owned(
                 INVALID_REQUEST_CODE,
@@ -417,7 +263,7 @@ impl EngineApiServer for RollupBoostServer {
     ) -> RpcResult<OpExecutionPayloadEnvelopeV4> {
         info!("received get_payload_v4");
 
-        match self.get_payload(payload_id, Version::V4).await? {
+        match self.get_payload(payload_id, PayloadVersion::V4).await? {
             OpExecutionPayloadEnvelope::V4(v4) => Ok(v4),
             OpExecutionPayloadEnvelope::V3(_) => Err(ErrorObject::owned(
                 INVALID_REQUEST_CODE,
@@ -457,104 +303,6 @@ impl EngineApiServer for RollupBoostServer {
     }
 }
 
-#[derive(Debug, Clone)]
-pub enum OpExecutionPayloadEnvelope {
-    V3(OpExecutionPayloadEnvelopeV3),
-    V4(OpExecutionPayloadEnvelopeV4),
-}
-
-impl OpExecutionPayloadEnvelope {
-    pub fn version(&self) -> Version {
-        match self {
-            OpExecutionPayloadEnvelope::V3(_) => Version::V3,
-            OpExecutionPayloadEnvelope::V4(_) => Version::V4,
-        }
-    }
-}
-
-impl From<OpExecutionPayloadEnvelope> for ExecutionPayload {
-    fn from(envelope: OpExecutionPayloadEnvelope) -> Self {
-        match envelope {
-            OpExecutionPayloadEnvelope::V3(v3) => ExecutionPayload::from(v3.execution_payload),
-            OpExecutionPayloadEnvelope::V4(v4) => {
-                ExecutionPayload::from(v4.execution_payload.payload_inner)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct NewPayloadV3 {
-    pub payload: ExecutionPayloadV3,
-    pub versioned_hashes: Vec<B256>,
-    pub parent_beacon_block_root: B256,
-}
-
-#[derive(Debug, Clone)]
-pub struct NewPayloadV4 {
-    pub payload: OpExecutionPayloadV4,
-    pub versioned_hashes: Vec<B256>,
-    pub parent_beacon_block_root: B256,
-    pub execution_requests: Vec<Bytes>,
-}
-
-#[derive(Debug, Clone)]
-pub enum NewPayload {
-    V3(NewPayloadV3),
-    V4(NewPayloadV4),
-}
-
-impl NewPayload {
-    pub fn version(&self) -> Version {
-        match self {
-            NewPayload::V3(_) => Version::V3,
-            NewPayload::V4(_) => Version::V4,
-        }
-    }
-}
-
-impl From<OpExecutionPayloadEnvelope> for NewPayload {
-    fn from(envelope: OpExecutionPayloadEnvelope) -> Self {
-        match envelope {
-            OpExecutionPayloadEnvelope::V3(v3) => NewPayload::V3(NewPayloadV3 {
-                payload: v3.execution_payload,
-                versioned_hashes: vec![],
-                parent_beacon_block_root: v3.parent_beacon_block_root,
-            }),
-            OpExecutionPayloadEnvelope::V4(v4) => NewPayload::V4(NewPayloadV4 {
-                payload: v4.execution_payload,
-                versioned_hashes: vec![],
-                parent_beacon_block_root: v4.parent_beacon_block_root,
-                execution_requests: v4.execution_requests,
-            }),
-        }
-    }
-}
-
-impl From<NewPayload> for ExecutionPayload {
-    fn from(new_payload: NewPayload) -> Self {
-        match new_payload {
-            NewPayload::V3(v3) => ExecutionPayload::from(v3.payload),
-            NewPayload::V4(v4) => ExecutionPayload::from(v4.payload.payload_inner),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum Version {
-    V3,
-    V4,
-}
-
-impl Version {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Version::V3 => "v3",
-            Version::V4 => "v4",
-        }
-    }
-}
-
 impl RollupBoostServer {
     async fn new_payload(&self, new_payload: NewPayload) -> RpcResult<PayloadStatus> {
         let execution_payload = ExecutionPayload::from(new_payload.clone());
@@ -588,7 +336,7 @@ impl RollupBoostServer {
     async fn get_payload(
         &self,
         payload_id: PayloadId,
-        version: Version,
+        version: PayloadVersion,
     ) -> RpcResult<OpExecutionPayloadEnvelope> {
         let l2_fut = self.l2_client.get_payload(payload_id, version);
 
@@ -657,6 +405,7 @@ impl RollupBoostServer {
                 if self.execution_mode().is_dry_run() {
                     (l2_payload, PayloadSource::L2)
                 } else {
+                    // NOTE: block selection policy
                     (payload, PayloadSource::Builder)
                 }
             } else {
